convert to newer gql autoads

+++ features/top_ads/top_ads_detail_sheet/src/main/java/com/tokopedia/topads/detail_sheet/viewmodel/TopAdsSheetViewModel.kt
@@ -89,7 +89,7 @@ class TopAdsSheetViewModel @Inject constructor(
     }
 
     fun getAutoAdsStatus(shopId: String, resources: Resources) {
-        val params = mapOf(ParamObject.SHOP_Id to shopId.toInt())
+        val params = mapOf(ParamObject.SHOP_Id to shopId)
         topAdsGetAutoAdsStatusUseCase.setGraphqlQuery(GraphqlHelper.loadRawString(resources, com.tokopedia.topads.common.R.raw.query_auto_ads_status))
         topAdsGetAutoAdsStatusUseCase.setTypeClass(TopAdsAutoAds.Response::class.java)
         topAdsGetAutoAdsStatusUseCase.setRequestParams(params)
